# Optimize encodeUint